Add files via upload
Initial commit. Also, the shortcut for the start.bat is experimental, if it doesn't work, just right click on the .bat and create it yourself. Than set admin permissions in 'advanced' and then it will start with admin permissions. Then you can choose a bunch of files, hit the tiny app's "open" button, then select where you want to link them to, and hit the tiny app's single button again. Couldn't be simpler. 

I purposefully made this extremely simply because I read through the entire script of any app I DL that requires admin permissions, and not wanting to create chores for people, I made this one without so much as a non-standardized font.